Move `Error` to `util`
