+++ readium/streamer/src/main/java/org/readium/r2/streamer/ParserAssetFactory.kt
@@ -10,14 +10,14 @@ import java.io.File
 import java.nio.charset.Charset
 import org.json.JSONObject
 import org.readium.r2.shared.asset.Asset
-import org.readium.r2.shared.error.MessageError
-import org.readium.r2.shared.error.ThrowableError
 import org.readium.r2.shared.publication.Manifest
 import org.readium.r2.shared.publication.Publication
 import org.readium.r2.shared.resource.Resource
 import org.readium.r2.shared.resource.ResourceContainer
 import org.readium.r2.shared.resource.RoutingContainer
 import org.readium.r2.shared.resource.StringResource
+import org.readium.r2.shared.util.MessageError
+import org.readium.r2.shared.util.ThrowableError
 import org.readium.r2.shared.util.Try
 import org.readium.r2.shared.util.getOrElse
 import org.readium.r2.shared.util.getOrThrow
