+++ readium/streamer/src/main/java/org/readium/r2/streamer/parser/PublicationParser.kt
@@ -6,11 +6,11 @@
 
 package org.readium.r2.streamer.parser
 
-import org.readium.r2.shared.error.MessageError
-import org.readium.r2.shared.error.ThrowableError
 import org.readium.r2.shared.publication.Publication
 import org.readium.r2.shared.resource.Container
 import org.readium.r2.shared.resource.Resource
+import org.readium.r2.shared.util.MessageError
+import org.readium.r2.shared.util.ThrowableError
 import org.readium.r2.shared.util.Try
 import org.readium.r2.shared.util.logging.WarningLogger
 import org.readium.r2.shared.util.mediatype.MediaType
@@ -46,18 +46,18 @@ public interface PublicationParser {
         warnings: WarningLogger? = null
     ): Try<Publication.Builder, Error>
 
-    public sealed class Error : org.readium.r2.shared.error.Error {
+    public sealed class Error : org.readium.r2.shared.util.Error {
 
         public class FormatNotSupported : Error() {
 
             override val message: String =
                 "Asset format not supported."
 
-            override val cause: org.readium.r2.shared.error.Error? =
+            override val cause: org.readium.r2.shared.util.Error? =
                 null
         }
 
-        public class ParsingFailed(override val cause: org.readium.r2.shared.error.Error?) : Error() {
+        public class ParsingFailed(override val cause: org.readium.r2.shared.util.Error?) : Error() {
 
             public constructor(message: String) : this(MessageError(message))
 
@@ -72,12 +72,12 @@ public interface PublicationParser {
             override val message: String =
                 "An IO error occurred."
 
-            override val cause: org.readium.r2.shared.error.Error =
+            override val cause: org.readium.r2.shared.util.Error =
                 ThrowableError(resourceError)
         }
 
         public class OutOfMemory(
-            override val cause: org.readium.r2.shared.error.Error?
+            override val cause: org.readium.r2.shared.util.Error?
         ) : Error() {
 
             override val message: String =
