+++ test-app/src/main/java/org/readium/r2/testapp/PublicationError.kt
@@ -9,8 +9,8 @@ package org.readium.r2.testapp
 import androidx.annotation.StringRes
 import org.readium.r2.shared.UserException
 import org.readium.r2.shared.asset.AssetRetriever
-import org.readium.r2.shared.error.Error
 import org.readium.r2.shared.publication.Publication
+import org.readium.r2.shared.util.Error
 
 sealed class PublicationError(@StringRes userMessageId: Int) : UserException(userMessageId) {
